requirements.txt
pandas
aiogram
requests
python-dotenv
httpx